refactor: medication service and controller

+++ src/main/java/hackatongrupo7/medicamentos_grupo7/medication/MedicationService.java
@@ -1,68 +1,75 @@
 package hackatongrupo7.medicamentos_grupo7.medication;
 
-import hackatongrupo7.medicamentos_grupo7.Notification.Notification;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.messaging.simp.SimpMessagingTemplate;
-import org.springframework.scheduling.annotation.Scheduled;
+import hackatongrupo7.medicamentos_grupo7.medication.dto.MedicationMapper;
+import hackatongrupo7.medicamentos_grupo7.medication.dto.MedicationRequest;
+import hackatongrupo7.medicamentos_grupo7.medication.dto.MedicationResponseDetails;
+import hackatongrupo7.medicamentos_grupo7.medication.dto.MedicationResponseSummary;
+import hackatongrupo7.medicamentos_grupo7.user.User;
+import hackatongrupo7.medicamentos_grupo7.utils.EntityUtil;
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.List;
 import java.util.Optional;
 
 @Service
+@RequiredArgsConstructor
+@Transactional
 public class MedicationService {
 
     private final MedicationRepository medicationRepository;
-    private final SimpMessagingTemplate messagingTemplate;
+    private final MedicationMapper medicationMapper;
+    private final EntityUtil mapperUtil;
 
-    @Autowired
-    public MedicationService(MedicationRepository medicationRepository,
-                             SimpMessagingTemplate messagingTemplate) {
-        this.medicationRepository = medicationRepository;
-        this.messagingTemplate = messagingTemplate;
+    @Transactional
+    public MedicationResponseDetails saveMedication(MedicationRequest medication, User user) {
+        Medication saveMedication =  medicationRepository.save(medicationMapper.toEntity(medication, user));
+        System.out.println(saveMedication.isActive());
+        System.out.println(saveMedication.isTaken());
+        return medicationMapper.fromEntityDetails(saveMedication);
     }
 
-    public Medication save(Medication medication) {
-        return medicationRepository.save(medication);
+    public List<MedicationResponseSummary> findAllMedicationsByUser(User user) {
+        List<Medication> medications = medicationRepository.findByUserId(user.getId());
+
+        if (medications.isEmpty()){throw  new RuntimeException("Empty list");}
+
+        return mapperUtil.mapEntitiesToDTOs(medications, medicationMapper::fromEntitySummary);
     }
 
-    public List<Medication> findAll() {
-        return medicationRepository.findAll();
+    public MedicationResponseDetails findById(Long id, User user) {
+        Medication medication =  medicationRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Id not found"));
+
+        if (medication.getUser().getUsername() != user.getUsername()){
+            throw new RuntimeException("Unauthorized");
+        };
+
+        return medicationMapper.fromEntityDetails(medication);
+
     }
 
-    public Optional<Medication> findById(Long id) {
-        return medicationRepository.findById(id);
-    }
+    public void delete(Long id, User user) {
+        Medication medication =  medicationRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Id not found"));
 
-    public void delete(Long id) {
+        if (medication.getUser().getUsername() != user.getUsername()){
+            throw new RuntimeException("Unauthorized");
+        };
+
         medicationRepository.deleteById(id);
     }
 
-    public Medication markAsTaken(Long id) {
+    public MedicationResponseSummary markAsTaken(Long id, User user) {
         Medication medication = medicationRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Medication not found"));
+
+        if (medication.getUser().getUsername() != user.getUsername()){
+            throw new RuntimeException("Unauthorized");
+        };
+
         medication.setTaken(true);
-        return medicationRepository.save(medication);
-    }
-
-    @Scheduled(cron = "0 * * * * *")
-    public void checkAndSendReminders() {
-        String nowHour = LocalDateTime.now().format(DateTimeFormatter.ofPattern("HH:mm"));
-        List<Medication> dueMedications = medicationRepository.findByHourAndNotTaken(nowHour);
-
-        for (Medication med : dueMedications) {
-            Notification dto = Notification.builder()
-                    .id(med.getId())
-                    .name(med.getName())
-                    .dose(med.getDose())
-                    .hour(med.getHour())
-                    .message("Is time to take " + med.getName() + " (" + med.getDose() + ")")
-                    .createdAt(LocalDateTime.now())
-                    .build();
-
-            messagingTemplate.convertAndSend("/topic/medication-reminders", dto);
-        }
+        return medicationMapper.fromEntitySummary(medicationRepository.save(medication));
     }
 }
